Refactor: S3Service -> S3Manager
